// src/app/services/swagger/model/createForumThreadMessage.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.0.1
 * Contact: bla@bla.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * A forum thread message on creation.
 */
export interface CreateForumThreadMessage { 
    /**
     * The id of the parent's forum thread.
     */
    idForumThread: number;
    /**
     * The message.
     */
    message?: string;
}