// src/app/services/swagger/model/forum.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.0.1
 * Contact: bla@bla.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Represents a single forum with it's access rules and threads.
 */
export interface Forum { 
    /**
     * The id of the forum.
     */
    idForum: number;
    /**
     * If set, the forum is for the defined alliance only.
     */
    idAlliance?: number;
    /**
     * If it has a role set, only the bearer can visit it.
     */
    role?: Forum.RoleEnum;
    /**
     * The forums title.
     */
    title: string;
    /**
     * The forums description.
     */
    description: string;
    /**
     * The forums creation timestamp.
     */
    createdAt: Date;
    /**
     * The id's of the forum's threads.
     */
    idForumThreads: Array<number>;
}
export namespace Forum {
    export type RoleEnum = 'USER' | 'ADMIN';
    export const RoleEnum = {
        USER: 'USER' as RoleEnum,
        ADMIN: 'ADMIN' as RoleEnum
    };
}